backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dpp-backend"
version = "0.1.0"
description = "FastAPI backend for the Digital Product Passport platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14,<2.0",
    "asyncpg>=0.29,<1.0",
    "fastapi>=0.110,<1.0",
    "pyjwt[crypto]>=2.9,<3.0",
    "uvicorn[standard]>=0.27,<1.0",
    "pydantic>=2.6,<3.0",
    "pydantic-settings>=2.2,<3.0",
    "psycopg[binary]>=3.2,<4.0",
    "sqlalchemy>=2.0,<3.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0,<9.0",
    "ruff>=0.9,<1.0",
    "httpx>=0.27,<1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q"
